Add support for Client Link-Layer Address Option (#373)
RFC6939 specifies a client option to relay the link-layer address of the
original client in the relayed message. This commit adjusts the OOB
plugin to first try and read the address from that option and falls back
to the old EUI64 parsing when that fails.

It should be noted that the EUI64 parsing does not work if the client has
IPv6 privacy extensions enabled as the server can not distinguish between
EUI64 and and a random IPv6 host address.

See: https://www.rfc-editor.org/rfc/rfc6939